chore: satisfy review
* move .desktop exec parser to its own file

fixup the auth logic
